requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
pandas-ta>=0.3.14b0

# Data & APIs
newsapi-python>=0.2.7
requests>=2.31.0
# Alpha Vantage uses requests (no separate package needed)

# LLM & NLP
# FinBERT (local, free)
transformers>=4.35.0
torch>=2.1.0
sentencepiece>=0.1.99
# Gemini 2.0 Flash (API, free tier available)
google-generativeai>=0.3.0

# Backtesting
vectorbt>=0.25.0
pyfolio-reloaded>=0.9.5

# Execution (Phase 4)
alpaca-trade-api>=3.0.0
ib_insync>=0.9.86
nest_asyncio>=1.5.6
scipy>=1.9.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
backoff>=2.2.1

# Optional: Dashboard (Phase 4)
streamlit>=1.28.0
